requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# LangChain
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19

# Modèles et validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Weaviate (préparé pour plus tard)
weaviate-client==4.4.0

# Configuration
python-dotenv==1.0.0

# Logging et monitoring
colorlog==6.8.2

